Move the configuration loading to a common module.
