+++ application.py
@@ -7,10 +7,6 @@ import calendar
 import functools
 
 try:
-    import configparser
-except ImportError:
-    import ConfigParser as configparser
-try:
     from xmlrpc import client
 except ImportError:
     import xmlrpclib as client
@@ -34,42 +30,7 @@ import tracxml
 import natural
 import trac_to_markdown
 
-
-def load_configuration():
-    defaults = {
-        "trac": {
-            "host": "",
-            "user": "",
-            "password": "",
-            "limit": "35",
-            "components": "",
-            "priorities": "lowest,low,normal,high,highest",
-            "types": "defect,enhancement,task",
-            "extra_fields": "",
-            "statuses": ""
-        },
-        "slack": {
-            "token": "",
-            "endpoint": "/trac-slack"
-        },
-        "logging": {
-            "file": "/var/log/trac-slack.log",
-            "level": "INFO",
-            "sentry": "",
-            "user": "www-data",
-        }
-    }
-    conf = configparser.ConfigParser()
-    # Load in default values.
-    for section, values in defaults.items():
-        conf.add_section(section)
-        for option, value in values.items():
-            conf.set(section, option, value)
-
-    if os.path.exists("/etc/trac-slack.conf"):
-        conf.read("/etc/trac-slack.conf")
-    return conf
-
+from core import load_configuration
 
 CONF = load_configuration()
 # This is the WSGI application that we are creating.
